dao-h2

+++ src/dao/IDao.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface IDao<Odontologo> {
     public List<Odontologo> buscarTodos();
-    public Odontologo buscar(Long id);
+    public Odontologo buscar(int id);
     public void guardar(Odontologo odontologo);
-    public void eliminar(Long id);
+    public void eliminar(int id);
 }
